requirements.txt
scrapy
awesome-slugify